X-ray counterpart and Gamma-ray counterpart function added
I've updated the source class with the x-ray and gamma-ray counterpart function. It finds the closest source in the corresponding catalogs and writes it's counterpart name and distance from the source. I've done a trial check for these functions by manually selecting a coordinate and checking if the functions are working correctly. You can find it in trialcheck.ipynb.